FastAPI_service/src/requirements.txt
aiohttp==3.8.3
aioredis==1.3.1
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
elastic-transport==8.4.0
elasticsearch==7.9.1
fastapi==0.89.0
filelock==3.9.0
flake8==6.0.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hiredis==2.1.1
identify==2.5.13
idna==3.4
mccabe==0.7.0
multidict==6.0.4
nodeenv==1.7.0
orjson==3.8.4
platformdirs==2.6.2
pre-commit==2.21.0
pycodestyle==2.10.0
pydantic==1.10.4
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==0.21.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
virtualenv==20.17.1
yarl==1.8.2
